+++ test/window/glfw_imp.hpp
@@ -110,28 +110,28 @@ namespace glfw
             glfwSetWindowTitle(window,str.data());
         }
 
-        float imp_getWindowPosX()
+        int imp_getWindowPosX()
         {
             int x;
             glfwGetWindowPos(window,&x, nullptr);
             return x;
         }
 
-        float imp_getWindowPosY()
+        int imp_getWindowPosY()
         {
             int y;
             glfwGetWindowPos(window,nullptr, &y);
             return y;
         }
 
-        float imp_getWindowSizeW()
+        int imp_getWindowSizeW()
         {
             int w;
             glfwGetWindowSize(window,&w,nullptr);
             return w;
         }
 
-        float imp_getWindowSizeH()
+        int imp_getWindowSizeH()
         {
             int h;
             glfwGetWindowSize(window,nullptr,&h);
